Ajuste Design Listagem de Tarefas

+++ views/index.php
@@ -17,39 +17,43 @@
 
 <div class='container-tarefas'>
 <?php
-    $chaves = array_keys($tarefasCategorizadas);
-    $cont = 0;
-    foreach($tarefasCategorizadas as $tarefas){
-        $chave = $chaves[$cont];
-        echo "
-            <div class='container-categoria'>
-                <div class='categoria-nome'>
-                    $chave 
-                </div>
-                <hr>
-            </div>
-        ";
-        foreach($tarefas as $tarefa){
-            if ($tarefa == false){
-                $cont++;
-                continue;
-            }else{
-                $caminho = APP;
-                $data = implode('/', array_reverse(explode('-', $tarefa['prazo'])));
-                echo "
-                    <div class='tarefa'>
-                        <div>
-                            <p class='nome-tarefa'>{$tarefa['descricao']}</p>
-                            <p class='prazo-tarefa'>$data</p>
-                        </div>
-                        <a class='butao deletar-tarefa' type='button' href='$caminho/tarefa/excluir/{$tarefa['id']}'>
-                            <p class='check-icon'>✓</p>
-                        </a>
+    if($tarefasCategorizadas != ""){
+        $chaves = array_keys($tarefasCategorizadas);
+        $cont = 0;
+        foreach($tarefasCategorizadas as $tarefas){
+            $chave = $chaves[$cont];
+            echo "
+                <div class='container-categoria'>
+                    <div class='categoria-nome'>
+                        $chave 
                     </div>
-                ";
-            }   
+                    <hr>
+                </div>
+            ";
+            foreach($tarefas as $tarefa){
+                if ($tarefa == false){
+                    $cont++;
+                    continue;
+                }else{
+                    $caminho = APP;
+                    $data = implode('/', array_reverse(explode('-', $tarefa['prazo'])));
+                    echo "
+                        <div class='tarefa'>
+                            <div>
+                                <p class='nome-tarefa'>{$tarefa['descricao']}</p>
+                                <p class='prazo-tarefa'>$data</p>
+                            </div>
+                            <a class='butao deletar-tarefa' type='button' href='$caminho/tarefa/excluir/{$tarefa['id']}'>
+                                <p class='check-icon'>✓</p>
+                            </a>
+                        </div>
+                    ";
+                }   
+            }
+            $cont++;
         }
-        $cont++;
+    } else{
+        echo "<p>Nenhuma categoria cadastrada</p>";
     }
 ?>
 </div>
